ProxyChannelStateMachineEmbeddedTest -> ProxyChannelStateMachineEndToEndTest
Clean up most sonar warnings as well.


rh-pre-commit.version: 2.0.1
rh-pre-commit.check-secrets: ENABLED
